refactor: move workflow to components
