+++ templates/components/multiagent/typescript/workflow/factory.ts
