+++ templates/components/multiagent/typescript/workflow/stream.ts
